add key for token and move function from helper to action
