+++ shared/helpers/bch.js
@@ -53,19 +53,6 @@ const estimateFeeValue = async ({ feeRate, inSatoshis, speed, address, txSize, f
   return finalFeeValue
 }
 
-const getTx = (txRaw) => {
-
-  return false
-}
-
-const getLinkToInfo = (tx) => {
-  if(!tx) {
-    return
-  }
-
-  return `https://www.blockchain.com/ru/btc/tx/${tx}`
-}
-
 const estimateFeeRate = async ({ speed = 'fast' } = {}) => {
   const link = config.feeRates.bch
   const defaultRate = constants.defaultFeeRates.bch.rate
@@ -102,7 +89,5 @@ export default {
   calculateTxSize,
   estimateFeeValue,
   estimateFeeRate,
-  network,
-  getTx,
-  getLinkToInfo
+  network
 }
