+++ shared/helpers/btc.js
@@ -57,20 +57,6 @@ const estimateFeeValue = async ({ feeRate, inSatoshis, speed, address, txSize, f
   return finalFeeValue
 }
 
-const getTx = (txRaw) => {
-
-  return txRaw.getId()
-}
-
-const getLinkToInfo = (tx) => {
-
-  if(!tx) {
-    return
-  }
-
-  return `${config.link.bitpay}/tx/${tx}`
-}
-
 const estimateFeeRate = async ({ speed = 'fast' } = {}) => {
   const link = config.feeRates.btc
   const defaultRate = constants.defaultFeeRates.btc.rate
@@ -107,7 +93,5 @@ export default {
   calculateTxSize,
   estimateFeeValue,
   estimateFeeRate,
-  network,
-  getTx,
-  getLinkToInfo
+  network
 }
