+++ shared/helpers/eth.js
@@ -14,21 +14,6 @@ const estimateFeeValue = async ({ method = 'send', speed } = {}) => {
   return feeValue
 }
 
-const getTx = (txRaw) => {
-
-  return txRaw.transactionHash
-}
-
-const getLinkToInfo = (tx) => {
-
-  if(!tx) {
-    return
-  }
-
-  return `https://etherscan.io/tx/${tx}`
-}
-
-
 const estimateGasPrice = async ({ speed = 'fast' } = {}) => {
   const link = config.feeRates.eth
   const defaultPrice = constants.defaultFeeRates.eth.price
@@ -61,7 +46,5 @@ const estimateGasPrice = async ({ speed = 'fast' } = {}) => {
 
 export default {
   estimateFeeValue,
-  estimateGasPrice,
-  getTx,
-  getLinkToInfo
+  estimateGasPrice
 }
