+++ shared/helpers/ethToken.js
@@ -17,20 +17,6 @@ const estimateFeeValue = async ({ method = 'send', speed } = {}) => {
 
   return feeValue
 }
-const getTx = (txRaw) => {
-
-  return txRaw.transactionHash
-}
-
-const getLinkToInfo = (tx) => {
-
-  if(!tx) {
-    return
-  }
-
-  return `https://etherscan.io/tx/${tx}`
-}
-
 
 const estimateGasPrice = ({ speed } = {}) => eth.estimateGasPrice({ speed })
 
@@ -38,7 +24,5 @@ export default {
   estimateFeeValue,
   estimateGasPrice,
   isEthToken,
-  isEthOrEthToken,
-  getTx,
-  getLinkToInfo
+  isEthOrEthToken
 }
