+++ shared/redux/actions/token.js
@@ -175,6 +175,20 @@ const fetchFees = async ({ gasPrice, gasLimit, speed } = {}) => {
   }
 }
 
+const getTx = (txRaw) => {
+
+  return txRaw.transactionHash
+}
+
+const getLinkToInfo = (tx) => {
+
+  if(!tx) {
+    return
+  }
+
+  return `https://etherscan.io/tx/${tx}`
+}
+
 const sendTransaction = ({ contract, method }, { args, params = {} } = {}, callback) =>
   new Promise(async (resolve, reject) => {
     const receipt = await contract.methods[method](...args).send(params)
@@ -254,6 +268,8 @@ export default {
   getBalance,
   getTransaction,
   send,
+  getTx,
+  getLinkToInfo,
   approve,
   setAllowanceForToken,
   fetchBalance,
